To solve this problem, we can start by initializing a counter variable to zero and another variable to represent the number of characters in the file, which we will initially set to 1 (since there is already one H in the file). We can then start a loop that will continue until the number of characters in the file is equal to n.
In each iteration of the loop, we can check if the current number of characters in the file is divisible by the number of characters we currently have (i.e., whether we can copy all the characters in the file and paste them multiple times to get more characters). If so, we can update the number of characters in the file by multiplying it by 2 and incrementing the counter variable by 2 (one operation for the copy and one for the paste). If not, we can simply increment the number of characters in the file by 1 and the counter variable by 1 (one operation for the paste).

At the end of the loop, if the number of characters in the file is equal to n, we can return the value of the counter variable. Otherwise, we return 0 to indicate that it was impossible to achieve the desired number of characters.