This implementation uses the fact that the only two operations we can perform are "Copy All" and "Paste". We start with a single H character, and we can only "Paste" a number of times equal to the number of H characters we have at the moment. Therefore, the number of H characters in the file must be a product of primes.
To find the prime factorization of n, we start by dividing n by 2 as many times as possible, then by 3 as many times as possible, then by 5, and so on, until we reach a prime factor. Each time we divide n by a prime factor, we add that factor to the number of operations.

If n is impossible to achieve (i.e., if n is not a product of primes), the function returns 0.